Working version.
This version reads the contents of a directory containing CORD-19 corpus .json files and pulls out the titles, abstracts, or bodies of the papers therein, or all three at once.  Just prints them--redirect to a file, or wait until the next version, where I'll make it output to a file.